crypto: Factor test utility functions
